+++ BeckendReportingSpringMongo/src/main/java/com/example/beckendreportingpengadaan/BidHistory/BidExchangeHistoryController.java
@@ -1,5 +1,8 @@
 package com.example.beckendreportingpengadaan.BidHistory;
 
+import com.example.beckendreportingpengadaan.BidHistory.DTO.BidExchangeHistoryRequestDTO;
+import com.example.beckendreportingpengadaan.BidHistory.DTO.BidExchangeHistoryResponseDTO;
+import com.example.beckendreportingpengadaan.BidItems.BidItemDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
