+++ BeckendReportingSpringMongo/src/main/java/com/example/beckendreportingpengadaan/BidHistory/BidExchangeHistoryService.java
@@ -1,5 +1,9 @@
 package com.example.beckendreportingpengadaan.BidHistory;
 
+import com.example.beckendreportingpengadaan.BidHistory.DTO.BidExchangeHistoryRequestDTO;
+import com.example.beckendreportingpengadaan.BidHistory.DTO.BidExchangeHistoryResponseDTO;
+import com.example.beckendreportingpengadaan.BidItems.BidItemDTO;
+
 import java.util.List;
 
 public interface BidExchangeHistoryService {
