+++ BeckendReportingSpringMongo/src/main/java/com/example/beckendreportingpengadaan/BidItems/BidItemRepository.java
@@ -1,4 +1,4 @@
-package com.example.beckendreportingpengadaan.BidHistory;
+package com.example.beckendreportingpengadaan.BidItems;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
