+++ BeckendReportingSpringMongo/src/main/java/com/example/beckendreportingpengadaan/BidItems/BidItemService.java
@@ -1,4 +1,4 @@
-package com.example.beckendreportingpengadaan.BidHistory;
+package com.example.beckendreportingpengadaan.BidItems;
 
 import java.util.List;
