+++ BeckendReportingSpringMongo/src/main/resources/application.properties
@@ -1,5 +1,5 @@
 
-spring.data.mongodb.uri=mongodb://36.93.137.4:27017/RSUDSAMRAT
+spring.data.mongodb.uri=mongodb://36.93.137.4:27018/RSUDSAMRAT
 server.port=8090
 logging.level.org.mongodb.driver=DEBUG
