+++ src/main/java/com/example/pengadaanrsudsamrat/order/ReportServiceDTO/BidItemDTO.java
@@ -16,4 +16,5 @@ public class BidItemDTO {
     private Double bidPriceChange;
     private Double totalBidPrice;
     private String status;
+    private String message;
 }
