+++ src/main/resources/application.properties
@@ -6,6 +6,7 @@ spring.datasource.password=samratrs_86
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.redis.host=36.93.137.4
 spring.redis.port=6379
+spring.zipkin.base-url=http://rsudsamrat.site:9411
 
 
 spring.jpa.hibernate.ddl-auto=update
